Refactor sql-based db engine classes
