+++ vial/db/sql/postgresql.py
@@ -1,9 +1,9 @@
 import psycopg2
 
-from .base import Engine
+from .base import SQL
 
 
-class Postgresql(Engine):
+class Postgresql(SQL):
     """
     """
     engine_name = 'PostgreSQL'
@@ -36,19 +36,6 @@ class Postgresql(Engine):
         )
         return self
 
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type is None:
-            try:
-                self._conn.commit()
-            except Exception as e:
-                self._conn.rollback()
-                raise e
-
-        else:
-            self._conn.rollback()
-
-        self._conn.close()
-
     @classmethod
     def _tables_query(cls):
         query = ' '.join(f"""
@@ -56,14 +43,6 @@ class Postgresql(Engine):
         WHERE table_schema = 'public'
         """.split())
         return query
-
-    def list_tables(self):
-        cur = self._conn.cursor()
-        cur.execute(Postgresql._tables_query())
-        results = [x[0] for x in cur.fetchall()]
-        self._conn.commit()
-        cur.close()
-        return results
 
     @classmethod
     def _create_query(cls, table_name, serial=False, fields=None):
@@ -82,16 +61,6 @@ class Postgresql(Engine):
         """.split())
         return query
 
-    def create(self, table_name, serial=False, fields=None):
-        cur = self._conn.cursor()
-        cur.execute(Postgresql._create_query(
-            table_name=table_name,
-            serial=serial,
-            fields=fields or {},
-        ))
-        self._conn.commit()
-        cur.close()
-
     @classmethod
     def _insert_query(cls, table_name, fields=None):
         query = ' '.join(f"""
@@ -99,13 +68,6 @@ class Postgresql(Engine):
         VALUES ({', '.join(["%s" for _ in fields])})
         """.split())
         return query
-
-    def insert(self, table_name, data=None):
-        cur = self._conn.cursor()
-        cur.execute(Postgresql._insert_query(table_name, fields=data.keys()), tuple(data.values()))
-        self._conn.commit()
-        cur.close()
-        return True
 
     @classmethod
     def _select_query(cls, table_name, fields=None, where=None, like=None, offset=None, limit=None):
@@ -122,21 +84,6 @@ class Postgresql(Engine):
         {f' LIMIT {limit}' if limit else ''};
         """.split())
         return query
-
-    def select(self, table_name, fields=None, where=None, like=None, offset=None, limit=None, one=False):
-        cur = self._conn.cursor()
-        cur.execute(Postgresql._select_query(
-            table_name=table_name,
-            fields=fields,
-            where=where,
-            like=like,
-            offset=offset,
-            limit=limit
-        ))
-        result = cur.fetchone() if one else cur.fetchall()
-        self._conn.commit()
-        cur.close()
-        return result
 
     @classmethod
     def _update_query(cls, table_name, where=None, updation=None):
@@ -163,10 +110,3 @@ class Postgresql(Engine):
         {f' WHERE {" AND ".join([f"{k}=(%s)" for k in where.keys()])}' if where else ''}
         """.split())
         return query
-
-    def delete(self, table_name, where=None):
-        cur = self._conn.cursor()
-        cur.execute(Postgresql._delete_query(table_name, where=where), tuple(where.values()))
-        self._conn.commit()
-        cur.close()
-        return True
