+++ vial/db/sql/sqlite.py
@@ -1,9 +1,9 @@
 import sqlite3
 
-from .base import Engine
+from .base import SQL
 
 
-class Sqlite(Engine):
+class Sqlite(SQL):
     """
     """
     engine_name = 'SQLite'
@@ -26,33 +26,12 @@ class Sqlite(Engine):
         self._conn = sqlite3.connect(self.dbname)
         return self
 
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type is None:
-            try:
-                self._conn.commit()
-            except Exception as e:
-                self._conn.rollback()
-                raise e
-
-        else:
-            self._conn.rollback()
-
-        self._conn.close()
-
     @classmethod
     def _tables_query(cls):
         query = ' '.join(f"""
         SELECT name FROM sqlite_master WHERE type='table'
         """.split())
         return query
-
-    def list_tables(self):
-        cur = self._conn.cursor()
-        cur.execute(Sqlite._tables_query())
-        results = [x[0] for x in cur.fetchall()]
-        self._conn.commit()
-        cur.close()
-        return results
 
     @classmethod
     def _create_query(cls, table_name, serial=False, fields=None):
@@ -71,16 +50,6 @@ class Sqlite(Engine):
         """.split())
         return query
 
-    def create(self, table_name, serial=False, fields=None):
-        cur = self._conn.cursor()
-        cur.execute(Sqlite._create_query(
-            table_name=table_name,
-            serial=serial,
-            fields=fields or {},
-        ))
-        self._conn.commit()
-        cur.close()
-
     @classmethod
     def _insert_query(cls, table_name, fields=None):
         fields = fields or ()
@@ -89,13 +58,6 @@ class Sqlite(Engine):
         VALUES ({', '.join(["?" for _ in fields])})
         """.split())
         return query
-
-    def insert(self, table_name, data=None):
-        cur = self._conn.cursor()
-        cur.execute(Sqlite._insert_query(table_name, fields=data.keys()), tuple(data.values()))
-        self._conn.commit()
-        cur.close()
-        return True
 
     @classmethod
     def _select_query(cls, table_name, fields=None, where=None, like=None, offset=None, limit=None):
@@ -112,21 +74,6 @@ class Sqlite(Engine):
         {f' OFFSET {offset}' if offset else ''};
         """.split())
         return query
-
-    def select(self, table_name, fields=None, where=None, like=None, offset=None, limit=None, one=False):
-        cur = self._conn.cursor()
-        cur.execute(Sqlite._select_query(
-            table_name=table_name,
-            fields=fields,
-            where=where,
-            like=like,
-            offset=offset,
-            limit=limit
-        ))
-        result = cur.fetchone() if one else cur.fetchall()
-        self._conn.commit()
-        cur.close()
-        return result
 
     @classmethod
     def _update_query(cls, table_name, where=None, updation=None):
@@ -153,10 +100,3 @@ class Sqlite(Engine):
         {f' WHERE {" AND ".join([f"{k}=(?)" for k in where.keys()])}' if where else ''}
         """.split())
         return query
-
-    def delete(self, table_name, where=None):
-        cur = self._conn.cursor()
-        cur.execute(Sqlite._delete_query(table_name, where=where), tuple(where.values()))
-        self._conn.commit()
-        cur.close()
-        return True
